iopc: fix compilation with GCC 12.2.1
iopc/iopc-parser.c:2116:9: error: use of bitwise '&' with boolean operands [-Werror,-Wbitwise-instead-of-logical,Semantic Issue]
    if ((!parse_function_desc(pp, IOP_F_RES,  fun, &arg_chunks))
        ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
iopc/iopc-parser.c:2116:9: note: cast one or both operands to int to silence this warning [Semantic Issue]

+++ iopc/iopc-parser.c
@@ -2113,8 +2113,8 @@ parse_function_stmt(iopc_parser_t *pp, qv_t(iopc_attr) *attrs,
     parse_function_desc(pp, IOP_F_ARGS, fun, &arg_chunks);
 
     /* XXX we use & to execute both function calls */
-    if ((!parse_function_desc(pp, IOP_F_RES,  fun, &arg_chunks))
-    &   (!parse_function_desc(pp, IOP_F_EXN,  fun, &arg_chunks)))
+    if ((int)(!parse_function_desc(pp, IOP_F_RES,  fun, &arg_chunks))
+    &   (int)(!parse_function_desc(pp, IOP_F_EXN,  fun, &arg_chunks)))
     {
         info_loc("function %s may be a candidate for async-ness",
                  fun->loc, fun->name);
